// src/app/page/master/jenis-produk/jenis-produk.interface.ts
import {EventEmitter} from '@angular/core';

export interface JenisProduk {
	namaJenisProduk?: string;//
	kdKelompokProduk?: number;//
	kdAccount?: number;
	kdJenisProduksiHead?: string;
	// kdDepartemen?: string;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}